Make sure the controller is not seen as a mouse and is accessible.
When an input device is seen as a mouse or keyboard and at the same time is
seen as a joystick, the input gets weird in most programs trying to access the
joystick part (ghost touches, etc.).

It is actually the opposite of this:
- https://github.com/ValveSoftware/steam-for-linux/issues/3384
- https://github.com/denilsonsa/udev-joystick-blacklist

Make sure the mouse part is disabled when plugging the controller in the system
and make sure it is accessible for games like in the Steam Controller rules.

With this patch, the controller works fine in Steam.